Move toLocalISOString to utils/time
