+++ src/features/forms/utils.ts
@@ -3,7 +3,7 @@ import { type HierarcicalObjectReference, type ObjectData, type ObjectId } from 
 
 import { HighlightCollection } from "contexts/highlightCollections";
 import { searchByPatterns } from "utils/search";
-import { sleep } from "utils/time";
+import { sleep, toLocalISOString } from "utils/time";
 
 import {
     DateTimeItem,
@@ -84,7 +84,7 @@ export async function idsToObjects({
 
             return acc;
         },
-        [[]] as string[][]
+        [[]] as string[][],
     );
 
     const concurrentRequests = 5;
@@ -108,7 +108,7 @@ export async function idsToObjects({
                         },
                     ],
                 }).catch(() => {});
-            })
+            }),
         );
 
         await sleep(1);
@@ -180,7 +180,7 @@ export async function mapGuidsToIds({
 
             return acc;
         },
-        [[]] as string[][]
+        [[]] as string[][],
     );
 
     const concurrentRequests = 5;
@@ -211,7 +211,7 @@ export async function mapGuidsToIds({
                     ],
                     full: true,
                 }).catch(() => {});
-            })
+            }),
         );
 
         await sleep(1);
@@ -502,35 +502,4 @@ export function determineHighlightCollection(form: Form): HighlightCollection {
         return HighlightCollection.FormsCompleted;
     }
     return HighlightCollection.FormsNew;
-}
-
-function toLocalISOString(date?: Date) {
-    if (!date) {
-        return;
-    }
-
-    let offset = "Z";
-
-    const tz = date.getTimezoneOffset();
-    if (tz !== 0) {
-        const tz_abs = Math.abs(tz);
-        const tz_hour = Math.floor(tz_abs / 60);
-        const tz_min = tz_abs % 60;
-        offset = `${tz > 0 ? "-" : "+"}${tz_hour.toString().padStart(2, "0")}:${tz_min.toString().padStart(2, "0")}`;
-    }
-
-    return (
-        date.getFullYear() +
-        "-" +
-        (date.getMonth() + 1).toString().padStart(2, "0") +
-        "-" +
-        date.getDate().toString().padStart(2, "0") +
-        "T" +
-        date.getHours().toString().padStart(2, "0") +
-        ":" +
-        date.getMinutes().toString().padStart(2, "0") +
-        ":" +
-        date.getSeconds().toString().padStart(2, "0") +
-        `.000000000${offset}`
-    );
 }
